Updated Post: Filter
While I still mute these terms, I didn’t want to be mistaken as someone
who does not care about these topics. That couldn’t be further from the
truth. I care, and want to get to the heart of the matter and avoid the
noise at the fringe of the discussions.